Create a separate license structure for Waymo-contributed 3D augmentations (#1598)
* Create a separate license structure for Waymo-contributed 3D augmentations

* Add  license to init

* Fix some imports

* Fix imports in tests

* Typo

* Update license
